改为 protobuf 作为传输格式

+++ utils.go
@@ -8,7 +8,7 @@ import (
 
 var _consistentHashing = crc32.ChecksumIEEE
 
-func GetConsistentHash(key []byte, partition int64) int64 {
+func GetConsistentHash(key []byte, partition int) int {
 	index := _consistentHashing(key)
-	return utils.JumpConsistentHash(uint64(index), partition)
+	return int(utils.JumpConsistentHash(uint64(index), int64(partition)))
 }
